api-sqlite/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
marshmallow==3.4.0
python-dotenv==0.11.0
SQLAlchemy==1.3.13
